CLeaned up some unused methods, made the alt_* files the primary versions of all files.
